+++ src/Main.java
@@ -5,7 +5,7 @@ public class Main {
         int weight = 98;
         BmiService service = new BmiService();
         int bmi;
-        bmi = (int) service.calculate(height, weight);
+        bmi = service.calculate(height, weight);
 
         System.out.println(bmi);
     }
